don't flatten directory as its confusing to look at fix readme example added vs code extensions.json move test directory around
