+++ package.json
@@ -17,15 +17,15 @@
   "bugs": {
     "url": "https://github.com/mrjono1/joi-to-typescript/issues"
   },
-  "main": "index.js",
-  "types": "index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
     "build": "tsc",
     "format": "prettier --write \"src/**/*.ts\" \"src/**/*.js\"",
     "lint": "eslint 'src/**'",
     "test": "jest --config jest.config.js",
     "coverage": "yarn test --coverage",
-    "pub": "tsc && cp package.json ./dist && cp README.md ./dist && cp LICENSE ./dist && cd ./dist && yarn publish"
+    "pub": "tsc && yarn publish"
   },
   "dependencies": {
     "joi": "^17.2.1"
