+++ src/__tests__/02/02.fromFile.ts
@@ -1,4 +1,4 @@
-import { convertFromDirectory } from '../index';
+import { convertFromDirectory } from '../../index';
 import { readFileSync } from 'fs';
 
 test('02.fromFiles', async () => {
@@ -20,7 +20,6 @@ test('02.fromFiles', async () => {
 
 export * from './FooBar';
 export * from './One';
-export * from './Readme';
 `
   );
   const oneContent = readFileSync(`${typeOutputDirectory}/One.ts`).toString();
@@ -65,33 +64,6 @@ export interface Foo {
    * Id
    */
   id: number;
-}
-`);
-
-  const readmeContent = readFileSync(`${typeOutputDirectory}/Readme.ts`).toString();
-
-  expect(readmeContent).toBe(`/**
- * This file was automatically generated by joi-to-typescript
- * Do not modify this file manually
- */
-
-export interface Job {
-  businessName: string;
-  jobTitle: string;
-}
-
-/**
- * A list of People
- */
-export type People = Person[];
-
-export interface Person {
-  firstName: string;
-  job?: Job;
-  /**
-   * Last Name
-   */
-  lastName: string;
 }
 `);
 });
