+++ src/__tests__/06/06.alternatives.ts
@@ -1,4 +1,4 @@
-import { convertFromDirectory } from '../index';
+import { convertFromDirectory } from '../../index';
 import { readFileSync } from 'fs';
 
 test('06.alternatives', async () => {
