+++ src/__tests__/readme/readme.ts
@@ -1,0 +1,40 @@
+import { convertFromDirectory } from '../../index';
+import { readFileSync } from 'fs';
+
+test('readme', async () => {
+  const typeOutputDirectory = './src/__tests__/readme/models';
+
+  const result = await convertFromDirectory({
+    schemaDirectory: './src/__tests__/readme/schemas',
+    typeOutputDirectory
+  });
+
+  expect(result).toBe(true);
+
+  const readmeContent = readFileSync(`${typeOutputDirectory}/Readme.ts`).toString();
+
+  expect(readmeContent).toBe(`/**
+ * This file was automatically generated by joi-to-typescript
+ * Do not modify this file manually
+ */
+
+export interface Job {
+  businessName: string;
+  jobTitle: string;
+}
+
+/**
+ * A list of People
+ */
+export type People = Person[];
+
+export interface Person {
+  firstName: string;
+  job?: Job;
+  /**
+   * Last Name
+   */
+  lastName: string;
+}
+`);
+});
